refactor: unify utils
