+++ src/tabs/tabs.ts
@@ -1,8 +1,8 @@
-import dom from '../behaviors/dom';
-import touch from '../behaviors/touch';
 import { SuperComponent, wxComponent, RelationsOptions, useId } from '../common/src/index';
 import props from './props';
 import config from '../common/config';
+import touch from '../mixins/touch';
+import { getRect } from '../common/utils';
 
 const { prefix } = config;
 const name = `${prefix}-tabs`;
@@ -15,7 +15,7 @@ enum Position {
 }
 @wxComponent()
 export default class Tabs extends SuperComponent {
-  behaviors = [dom, touch];
+  behaviors = [touch];
 
   externalClasses = [`${prefix}-class`, `${prefix}-class-item`, `${prefix}-class-active`, `${prefix}-class-track`];
 
@@ -88,8 +88,8 @@ export default class Tabs extends SuperComponent {
     });
 
     this.adjustPlacement();
-    this.gettingBoundingClientRect(`.${name}`).then((res: any) => {
-      this.containerWidth = res.width;
+    getRect(this, `.${name}`).then((rect) => {
+      this.containerWidth = rect.width;
     });
   }
 
@@ -162,7 +162,7 @@ export default class Tabs extends SuperComponent {
         resolve(this.trackWidth);
         return;
       }
-      this.gettingBoundingClientRect(`.${prefix}-tabs__track`).then((res) => {
+      getRect(this, `.${prefix}-tabs__track`).then((res) => {
         if (res) {
           this.trackWidth = res.width;
           resolve(this.trackWidth);
@@ -179,18 +179,18 @@ export default class Tabs extends SuperComponent {
     if (currentIndex <= -1) return;
 
     try {
-      const res = await this.gettingBoundingClientRect(`.${prefix}-tabs__item`, true);
+      const res = await getRect(this, `.${prefix}-tabs__item`, true);
       const rect = res[currentIndex];
       if (!rect) return;
       let count = 0;
       let distance = 0;
-      // eslint-disable-next-line no-restricted-syntax
-      for (const item of res) {
+
+      res.forEach((item) => {
         if (count < currentIndex) {
           distance += isScrollX ? item.width : item.height;
           count += 1;
         }
-      }
+      });
 
       if (this.containerWidth) {
         const offset = this.calcScrollOffset(this.containerWidth, rect.left, rect.width, this.data.offset);
